@@ -1,10 +1,14 @@
-const { defineConfig } = require("cypress");
+const { defineConfig } = require('cypress')
 
 module.exports = defineConfig({
+ 
   e2e: {
     setupNodeEvents(on, config) {
       // implement node event listeners here
     },
-    baseUrl: 'http://qamid.tmweb.ru',
+    retries: {
+      runMode: 1,
+      openMode: 1,
+    },
   },
 });
